refactor: move logger to src/infra/
