+++ src/handlers/close-task.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, test } from "bun:test";
 import { CoderAPIError } from "../coder-client";
-import { TestLogger } from "../logger";
+import { TestLogger } from "../infra/logger";
 import type { HandlerConfig } from "../schemas";
 import {
 	MockCoderClient,
