+++ src/handlers/create-task.ts
@@ -1,7 +1,7 @@
 import type { CoderClient } from "../coder-client";
 import { TaskNameSchema } from "../coder-client";
 import type { GitHubClient } from "../github-client";
-import type { Logger } from "../logger";
+import type { Logger } from "../infra/logger";
 import type { ActionOutputs, HandlerConfig } from "../schemas";
 import { generateTaskName } from "../task-utils";
