+++ src/handlers/failed-check.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, test } from "bun:test";
-import { TestLogger } from "../logger";
+import { TestLogger } from "../infra/logger";
 import type { HandlerConfig } from "../schemas";
 import {
 	MockCoderClient,
