+++ src/handlers/pr-comment.ts
@@ -1,6 +1,6 @@
 import type { CoderClient } from "../coder-client";
 import type { GitHubClient } from "../github-client";
-import type { Logger } from "../logger";
+import type { Logger } from "../infra/logger";
 import { formatPRCommentMessage } from "../messages";
 import type { ActionOutputs, HandlerConfig } from "../schemas";
 import {
