+++ src/task-utils.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, mock, test } from "bun:test";
 import type { CoderClient } from "./coder-client";
 import { CoderAPIError } from "./coder-client";
-import { TestLogger } from "./logger";
+import { TestLogger } from "./infra/logger";
 import { MockCoderClient, mockTask } from "./test-helpers";
 import {
 	generateTaskName,
